各テストコード

+++ app/controllers/comments_controller.rb
@@ -1,13 +1,26 @@
 class CommentsController < ApplicationController
   def create
-    @comment = Comment.create(comment_params)
-    redirect_to post_path(@comment.post.id)
+    @comment = Comment.new(comment_params)
+    if @comment.valid?
+      @comment.save
+      redirect_to post_path(@comment.post.id)
+    else
+      @post = @comment.post
+      @comments = @post.comments
+      @q = Post.ransack(params[:q])
+      render "posts/show"
+    end
   end
 
   def destroy
-    comment = Comment.find(params[:id])
-    comment.destroy
-    redirect_to post_path(comment.post.id)
+    @post = Post.find(params[:post_id])
+    comment = @post.comments.find(params[:id])
+    if current_user.id == comment.user.id
+      comment.destroy
+      redirect_to post_path(comment.post.id)
+    else
+      render "posts/show"
+    end
   end
 
   private
